Move popover.js to assets and fix webpack
