+++ meinberlin/assets/js/app.js
@@ -28,7 +28,7 @@ var ReactPolls = require('../../apps/polls/assets/react_polls.jsx')
 
 var relativeTimestamps = require('../../apps/actions/assets/timestamps.js')
 var mapAddress = require('./map-address.js')
-var remarkpopover = require('../../apps/moderatorremark/static/js/popover.js')
+var remarkpopover = require('../../apps/moderatorremark/assets/popover.js')
 var dynamicFields = require('../../apps/contrib/assets/dynamic_fields.js')
 
 // This function is overwritten with custom behavior in embed.js.
