+++ webpack.common.js
@@ -17,7 +17,6 @@ module.exports = {
       'jquery.scrollintoview',
       'js-cookie',
       'react',
-      'bootstrap',
       'immutability-helper',
       'react-dom',
       'react-flip-move',
